requirements.txt
asgiref==3.2.10
astroid==2.4.2
blessings==1.7
bpython==0.20
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
curtsies==0.3.4
Django==3.1.2
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-filter==2.4.0
django-multiselectfield==0.1.12
django-select2==7.4.2
django-widget-tweaks==1.4.8
dropbox==10.8.0
greenlet==0.4.17
idna==2.10
isort==5.5.2
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.22
mysqlclient==2.0.1
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
Pygments==2.7.2
pylint==2.6.0
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-decouple==3.3
python-dotenv==0.15.0
python-http-client==3.3.1
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
wrapt==1.12.1